Removes use of deprecated PType.numeric()

+++ partiql-planner/src/main/kotlin/org/partiql/planner/internal/transforms/RelConverter.kt
@@ -751,5 +751,5 @@ internal object RelConverter {
     private val STRUCT: CompilerType = CompilerType(PType.struct())
     private val BAG: CompilerType = CompilerType(PType.bag())
     private val LIST: CompilerType = CompilerType(PType.array())
-    private val INT: CompilerType = CompilerType(PType.numeric())
+    private val INT: CompilerType = CompilerType(PType.numeric(38, 0))
 }
