+++ partiql-spi/src/main/kotlin/org/partiql/value/PartiQLValueType.kt
@@ -84,7 +84,7 @@ public enum class PartiQLValueType {
             LIST -> PType.array()
             BAG -> PType.bag()
             FLOAT32 -> PType.real()
-            INT -> PType.numeric()
+            INT -> PType.numeric(38, 0)
             INT64 -> PType.bigint()
             INT32 -> PType.integer()
             INT16 -> PType.smallint()
